requirements.txt
openai==1.54.3
httpx==0.27.0
chromadb==0.5.23
sentence-transformers==3.3.1
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
wikipedia==1.4.0
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2
tiktoken==0.8.0